+++ dispatcher/rpcservice/idgenrator.go
@@ -2,7 +2,7 @@ package rpcservice
 
 // IDGenerator is a time based id generator.
 type IDGenerator struct {
-	id int
+	id uint64
 }
 
 // NewIDGenerator creates a new id generator.
@@ -11,12 +11,12 @@ func NewIDGenerator() *IDGenerator {
 }
 
 // SetID sets the id.
-func (g *IDGenerator) SetID(id int) {
+func (g *IDGenerator) SetID(id uint64) {
 	g.id = id
 }
 
 // ID return the next id.
-func (g *IDGenerator) ID() int {
+func (g *IDGenerator) ID() uint64 {
 	g.id++
 	return g.id
 }
